minor bug fixing around toml parsing

+++ src/minknow/Data.hpp
@@ -79,8 +79,8 @@ namespace readuntil
             bool runs = false;
             bool unblock_all = false;
             uint8_t actionBatchSize = 50;
-            uint8_t minChannel = 1;
-            uint8_t maxChannel = 512;
+            uint16_t minChannel = 1;
+            uint16_t maxChannel = 512;
             
             void addUnblockAction(GetLiveReadsRequest_Actions* actionList, uint32_t channelNr, uint32_t readNr, const double unblock_duration);
             void addStopReceivingDataAction(GetLiveReadsRequest_Actions* actionList, uint32_t channelNr, uint32_t readNr);
@@ -151,7 +151,7 @@ namespace readuntil
                 unblock_all = unblock;
             }
 
-            inline void setChannels(const uint8_t minC, const uint8_t maxC)
+            inline void setChannels(const uint16_t minC, const uint16_t maxC)
             {
                 minChannel = minC;
                 maxChannel = maxC;
